+++ libs/estdlib/src/gen_udp_socket.erl
@@ -402,7 +402,22 @@ handle_passive_recvfrom(State, Ref, From, Length, _Timeout) ->
 
 %% @private
 call(Pid, Request) ->
-    gen_server:call(Pid, Request, infinity).
+    try
+        gen_server:call(Pid, Request, infinity)
+    catch
+        exit:{Reason, _} when
+            Reason =:= normal;
+            Reason =:= noproc;
+            Reason =:= shutdown
+        ->
+            dead_reply(Request);
+        exit:{{shutdown, _}, _} ->
+            dead_reply(Request)
+    end.
+
+%% @private
+dead_reply(close) -> ok;
+dead_reply(_) -> {error, closed}.
 
 %% @private
 maybe_encode_binary(Options, Data) ->